% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse-comments.R
\name{collapse_comments}
\alias{collapse_comments}
\title{Collapse Comments}
\usage{
collapse_comments(...)
}
\arguments{
\item{...}{objects to be collapsed into a string.}
}
\description{
Collapse comments by concatenating, coercing to string
and then collapsing into a single string (character scalar)
using the '. ' separator.
}
\examples{
collapse_comments(c("Saw fish"), character(0), "Nice. .", NA_character_)
}
