% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates.R
\name{duplicates}
\alias{duplicates}
\title{Find duplicate rows in a data frame}
\usage{
duplicates(.data, ...)
}
\arguments{
\item{.data}{A data.frame.}

\item{...}{Optional column names to use when searching for duplicate rows in specific columns.}
}
\value{
A tibble containing only duplicate rows.
}
\description{
Finds duplicate rows within a data frame. Column names can optionally be
provided. If column names aren't provided, the entire data frame (all
columns) will be searched for duplicate rows. If column names are provided,
only the specified columns will be searched for duplicate rows.
}
