% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str-crush.R
\name{str_crush}
\alias{str_crush}
\title{Remove whitespace from a string}
\usage{
str_crush(string)
}
\arguments{
\item{string}{A character vector.}
}
\description{
\code{str_crush()}, removes all whitespace from a string,
is the logical extension to \code{\link[stringr:str_trim]{stringr::str_trim()}} and \code{\link[stringr:str_trim]{stringr::str_squish()}}.
}
\details{
\code{str_crush()} is considered \href{https://github.com/tidyverse/stringr/pull/338}{too specialized} to be part of stringr.
}
\examples{
str_crush("  String with trailing,  middle, and leading white space\t")
}
