% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str-to-snake-case.R
\name{str_to_snake_case}
\alias{str_to_snake_case}
\title{Converts strings to Snake Case}
\usage{
str_to_snake_case(x)
}
\arguments{
\item{x}{input string or mutliple strings to be converted to snake case}
}
\value{
string or strings converted to snake_case
}
\description{
Converts strings to Snake Case
}
\examples{
input <- "string of words"
output <- str_to_snake_case(input)
print(output)

input <- "StringOfWords"
output <- str_to_snake_case(input)
print(output)

input <- "s!t$ring of \%char^&act*ers"
output <- str_to_snake_case(input)
print(output)

input <- c("multiples of strings", "strings in a multiples", "many strings")
output <- str_to_snake_case(input)
print(output)

}
