% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyplus-package.R
\docType{package}
\name{tidyplus-package}
\alias{tidyplus}
\alias{tidyplus-package}
\title{tidyplus: Additional 'Tidyverse' Functions}
\description{
Provides functions that complement those in the 'tidyverse'.
}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID})

Other contributors:
\itemize{
  \item Poisson Consulting [copyright holder, funder]
}

}
\keyword{internal}
