% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates.R
\name{duplicates}
\alias{duplicates}
\title{Find duplicate rows in a data frame}
\usage{
duplicates(x, cols = colnames(x))
}
\arguments{
\item{x}{A data.frame.}

\item{cols}{A character vector of the column names.}
}
\value{
A tibble containing only duplicate rows.
}
\description{
Finds duplicate rows within a data frame. Column names can optionally be
provided. If column names aren't provided, the entire data frame (all
columns) will be searched for duplicate rows. If column names are provided,
only the specified columns will be searched for duplicate rows.
}
