% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates.R
\name{duplicates}
\alias{duplicates}
\title{Duplicates}
\usage{
duplicates(x, cols = colnames(x))
}
\arguments{
\item{x}{A data.frame.}

\item{cols}{A character vector of the column names.}
}
\value{
A tibble with just the duplicate rows included.
}
\description{
Duplicates
}
