% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyplus-package.R
\docType{package}
\name{tidyplus-package}
\alias{tidyplus}
\alias{tidyplus-package}
\title{tidyplus: Additional 'tidyverse' Functions}
\description{
Provides functions such as str_crush(), add_missing_column(), coalesce_data() and drop_na_all() that complement 'tidyverse' functionality or functions that provide alternative behaviors such as if_else2() and str_detect2().
}
\seealso{
Useful links:
\itemize{
  \item \url{https://poissonconsulting.github.io/tidyplus/}
  \item \url{https://github.com/poissonconsulting/tidyplus}
  \item Report bugs at \url{https://github.com/poissonconsulting/tidyplus/issues}
}

}
\author{
\strong{Maintainer}: Ayla Pearson \email{ayla@poissonconsulting.ca} (\href{https://orcid.org/0000-0001-7388-1222}{ORCID})

Authors:
\itemize{
  \item Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID})
}

Other contributors:
\itemize{
  \item Duncan Kennedy \email{duncan@poissonconsulting.ca} (\href{https://orcid.org/0009-0001-4880-5751}{ORCID}) [contributor]
  \item Poisson Consulting [copyright holder, funder]
}

}
\keyword{internal}
